Rating and account CRUD changes

+++ src/main/java/com/soprasteria/devopsassesmenttool/controller/CategoryController.java
@@ -39,18 +39,18 @@ public class CategoryController {
 	}
 
 	@RequestMapping(value = "/category/{categoryId}", method = RequestMethod.GET)
-	public Optional<Category> getCategoryById(@PathVariable(value = "categoryId") Integer categoryId) {
+	public Optional<Category> getCategoryById(@PathVariable(value = "categoryId") Long categoryId) {
 		return categoryService.getCategoryById(categoryId);
 	}
 
 	@RequestMapping(value = "/category", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public Category updateCategory(@PathVariable(value = "categoryId") Integer categoryId,
+	public Category updateCategory(@PathVariable(value = "categoryId") Long categoryId,
 			@RequestBody Category category) {
 		return categoryService.updateCategoryById(categoryId, category);
 	}
 
 	@RequestMapping(value = "/category/{categoryId}", method = RequestMethod.DELETE)
-	public ResponseEntity<Object> deleteCategoryById(@PathVariable(value = "categoryId") Integer categoryId) {
+	public ResponseEntity<Object> deleteCategoryById(@PathVariable(value = "categoryId") Long categoryId) {
 		return categoryService.deleteCategoryById(categoryId);
 	}
